pyproject.toml
[tool.poetry]
name = "flight-tracker"
version = "0.0.0"
description = "simple flight tracker for MSFS"
authors = ["Alex Chesters <10558792+AlexChesters@users.noreply.github.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
simconnect = {git = "https://github.com/AlexChesters/Python-SimConnect.git"}
typing-extensions = "^4.2.0"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
pylint = "^2.13.8"
pyinstaller = "^5.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
